refactor: 이미지 생성 책임 전가
